+++ app/analyze/ComparisonBarChart.tsx
