+++ app/analyze/MPTButtonsTitles.tsx
@@ -1,0 +1,44 @@
+export const mainButtonTitle = (page_number: number) => {
+  if (page_number == 1) {
+    return 'Modern Portfolio Theory';
+  }
+  if (page_number == 2) {
+    return 'Show my Portfolio';
+  }
+  if (page_number == 3) {
+    return 'Optimize my Portfolio';
+  }
+  if (page_number == 4) {
+    return 'Start from the beginning';
+  }
+};
+
+export const secondButtonTitle = (page_number: number) => {
+  if (page_number == 1) {
+    return 'Efficient Frontier (Coming soon)';
+  }
+  if (page_number == 2) {
+    return 'Explain Modern Portfolio Theory in more detail';
+  }
+  if (page_number == 3) {
+    return 'Show my Portfolio in more detail';
+  }
+  if (page_number == 4) {
+    return 'Explain portfolio optimization in more detail';
+  }
+};
+
+export const thirdButtonTitle = (page_number: number) => {
+  if (page_number == 1) {
+    return 'Simulate crashes (Coming soon)';
+  }
+  if (page_number == 2) {
+    return 'Explain Diversification';
+  }
+  if (page_number == 3) {
+    return 'Explain Risk managment';
+  }
+  if (page_number == 4) {
+    return 'Recalculate';
+  }
+};
