+++ app/analyze/SingleBarChart.tsx
