+++ app/analyze/index.tsx
@@ -1,12 +1,15 @@
 import React, { useState } from 'react';
 
 import { StyleSheet } from 'react-native';
+import { LogBox } from 'react-native';
 
 import { SafeAreaView } from '@/components/base/SafeAreaView';
 
-import ExplainView from './explainview';
-import NavigationButtons from './navigationbuttons';
-import PlotsView from './plotsview';
+import ExplainView from './Explain';
+import MPTButtons from './MPTButtons';
+import PlotsView from './Plots';
+
+LogBox.ignoreLogs(['Image source "null" doesn\'t exist']);
 
 export default function AnalyzeScreen() {
   // const [depotID, setDepotID] = useState('671a81f4a36d31f8e9d4e820'); // with nvidia and etfs
@@ -14,8 +17,10 @@ export default function AnalyzeScreen() {
   // const [depotID, setDepotID] = useState('61b0f5bc48801f001b86425f'); // user depot 2
   const [depotID, setDepotID] = useState('672a47e0bae468d6209a8bcc'); // demo depot
   const [portfolioID, setPortfolioID] = useState('');
+
   const [depotData, setDepotData] = useState({});
   const [mptData, setMPTData] = useState({});
+
   const [targetReturn, setTargetReturn] = useState(0.07);
   const [returnRange, setReturnRange] = useState<number[]>([0.01, 1]);
 
@@ -26,7 +31,7 @@ export default function AnalyzeScreen() {
 
   return (
     <SafeAreaView style={styles.container}>
-      <NavigationButtons
+      <MPTButtons
         depotID={depotID}
         portfolioID={portfolioID}
         setPortfolioID={setPortfolioID}
